requirements.txt
django-allauth[mfa,saml,socialaccount,steam]>=65.5.0
python-dotenv